# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "My Doc API",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/products": {
            "get": {
                "tags": [
                    "Product"
                ],
                "summary": "Список",
                "operationId": "10c7a540a415989276bd85db4e56e636",
                "responses": {
                    "200": {
                        "description": "Ok"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Product"
                ],
                "summary": "Создание",
                "operationId": "9936e8ee6c69d9ab9561b82b2cd75ee8",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Дег"
                                    },
                                    "id_sklad": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "selling_price": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 200
                                    },
                                    "id_type_product": {
                                        "type": "integer",
                                        "example": 2
                                    },
                                    "quantity": {
                                        "type": "integer",
                                        "example": 10
                                    },
                                    "status": {
                                        "type": "boolean",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Ok"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/products/{product}": {
            "get": {
                "tags": [
                    "Product"
                ],
                "summary": "Список по ID",
                "operationId": "174c6ca46e998dd579c0000f00a04999",
                "parameters": [
                    {
                        "name": "product",
                        "in": "path",
                        "description": "ID product",
                        "required": true,
                        "example": 1
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Ok"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Product"
                ],
                "summary": "Удаление",
                "operationId": "2f46a07936a5567af72b0feaa9f603d6",
                "parameters": [
                    {
                        "name": "product",
                        "in": "path",
                        "description": "ID product",
                        "required": true,
                        "example": 1
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Ok",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "done"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "patch": {
                "tags": [
                    "Product"
                ],
                "summary": "Обновление",
                "operationId": "961103789b2bea02d87d9ce9794817c1",
                "parameters": [
                    {
                        "name": "product",
                        "in": "path",
                        "description": "ID product",
                        "required": true,
                        "example": 2
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Дег"
                                    },
                                    "id_sklad": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "selling_price": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 200
                                    },
                                    "id_type_product": {
                                        "type": "integer",
                                        "example": 2
                                    },
                                    "quantity": {
                                        "type": "integer",
                                        "example": 10
                                    },
                                    "status": {
                                        "type": "boolean",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Ok"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/rawmaterials": {
            "get": {
                "tags": [
                    "RawMaterial"
                ],
                "summary": "Список",
                "operationId": "3375e840b5dbd458d776de0b9014731a",
                "responses": {
                    "200": {
                        "description": "Ok"
                    }
                }
            },
            "post": {
                "tags": [
                    "RawMaterial"
                ],
                "summary": "Создание",
                "operationId": "1f339bce2330c08f54698106bf0cbb09",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Алюминия"
                                    },
                                    "unit": {
                                        "type": "string",
                                        "example": "кг"
                                    },
                                    "quantity": {
                                        "type": "integer",
                                        "example": 6000
                                    },
                                    "purchase_price": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 20
                                    },
                                    "units_of_measurement": {
                                        "type": "string",
                                        "example": "сомони"
                                    },
                                    "id_supplier": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "date": {
                                        "type": "string",
                                        "format": "date-time",
                                        "example": "2024-02-14 15:00:15"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Описание"
                                    },
                                    "status": {
                                        "type": "boolean",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Ok"
                    }
                }
            }
        },
        "/api/rawmaterials/{rawmaterial}": {
            "get": {
                "tags": [
                    "RawMaterial"
                ],
                "summary": "Список по ID",
                "operationId": "0ac18f762c5f3a0c653ac3f57404f82c",
                "parameters": [
                    {
                        "name": "rawmaterial",
                        "in": "path",
                        "description": "ID сырье",
                        "required": true,
                        "example": 1
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Ok"
                    }
                }
            },
            "delete": {
                "tags": [
                    "RawMaterial"
                ],
                "summary": "Удаление",
                "operationId": "db6b64b024eed57e3aebb1c95b20527f",
                "parameters": [
                    {
                        "name": "rawmaterial",
                        "in": "path",
                        "description": "ID сырье",
                        "required": true,
                        "example": 1
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Ok",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "done"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "patch": {
                "tags": [
                    "RawMaterial"
                ],
                "summary": "Обновление",
                "operationId": "0f5bf7a401debd6aaf0ea69e58306d82",
                "parameters": [
                    {
                        "name": "rawmaterial",
                        "in": "path",
                        "description": "ID сырье",
                        "required": true,
                        "example": 1
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Алюминия"
                                    },
                                    "unit": {
                                        "type": "string",
                                        "example": "кг"
                                    },
                                    "quantity": {
                                        "type": "integer",
                                        "example": 6000
                                    },
                                    "purchase_price": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 20
                                    },
                                    "units_of_measurement": {
                                        "type": "string",
                                        "example": "сомони"
                                    },
                                    "id_supplier": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "date": {
                                        "type": "string",
                                        "format": "date-time",
                                        "example": "2024-02-14 15:00:15"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Описание"
                                    },
                                    "status": {
                                        "type": "boolean",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Ok"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    }
}